Re-Organize all API Handlers
